# open-api/openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/admin-signup": {
      "post": {
        "operationId": "createFirstAdmin",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentialDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/diary": {
      "get": {
        "operationId": "getAllDiaries",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "Diary"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "Diary"
        ]
      }
    },
    "/diary/{date}": {
      "delete": {
        "operationId": "deleteByDate",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "Diary"
        ]
      },
      "get": {
        "operationId": "getDiary",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "Diary"
        ]
      },
      "patch": {
        "operationId": "editDiary",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "path",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "Diary"
        ]
      }
    },
    "/server-info/config": {
      "get": {
        "operationId": "getServerConfig",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerConfigDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Server Info"
        ]
      }
    },
    "/user/me": {
      "get": {
        "operationId": "getMyUser",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    }
  },
  "info": {
    "title": "PRM",
    "description": "PRM's Api",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "": {
        "properties": {},
        "type": "object"
      },
      "AuthCredentialDto": {
        "properties": {
          "email": {
            "allOf": [
              {
                "$ref": "#/components/schemas/"
              }
            ],
            "example": "test@example.com"
          },
          "password": {
            "allOf": [
              {
                "$ref": "#/components/schemas/"
              }
            ],
            "example": "thispasswordisnotsafe"
          },
          "username": {
            "example": "testuser",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "CreateUserDto": {
        "properties": {
          "email": {
            "example": "test@example.com",
            "type": "string"
          },
          "name": {
            "allOf": [
              {
                "$ref": "#/components/schemas/"
              }
            ],
            "example": "John Doe"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "format": "username",
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "password",
          "username"
        ],
        "type": "object"
      },
      "DiaryCreateDto": {
        "properties": {
          "content": {
            "example": "Today...",
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "rating": {
            "example": 5,
            "maximum": 10,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "content",
          "date"
        ],
        "type": "object"
      },
      "DiaryEditDto": {
        "properties": {
          "content": {
            "allOf": [
              {
                "$ref": "#/components/schemas/"
              }
            ],
            "example": "Today..."
          },
          "rating": {
            "example": 5,
            "maximum": 10,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "content"
        ],
        "type": "object"
      },
      "ServerConfigDto": {
        "properties": {
          "isOnboarded": {
            "type": "boolean"
          }
        },
        "required": [
          "isOnboarded"
        ],
        "type": "object"
      },
      "UserResponseDto": {
        "properties": {
          "email": {
            "example": "test@example.com",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "id",
          "isAdmin",
          "name"
        ],
        "type": "object"
      }
    }
  }
}